requirements.txt
absl-py==0.9.0
appdirs==1.4.4
attrs==19.3.0
backcall==0.1.0
black==19.10b0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
future==0.18.2
google-auth==1.15.0
google-auth-oauthlib==0.4.1
grpcio==1.29.0
idna==2.9
importlib-metadata==1.6.0
ipdb==0.13.2
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.2.1
numpy==1.18.4
oauthlib==3.1.0
opencv-python-headless==4.2.0.34
parso==0.7.0
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pkg-resources==0.0.0
prompt-toolkit==3.0.5
protobuf==3.12.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-lightning==0.7.6
PyYAML==5.3.1
regex==2020.5.14
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
toml==0.10.1
torch==1.5.0
torchvision==0.6.0
tqdm==4.46.0
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.9
wcwidth==0.1.9
Werkzeug==1.0.1
zipp==3.1.0
